# ed25519: refactor