… contrib/codegen/gen_wycheproofs.py
… contrib/ed25519/dalek_target/Cargo.toml
… src/ballet/ed25519/Local.mk
… src/ballet/ed25519/avx/fd_ed25519_fe.c
… src/ballet/ed25519/avx/fd_ed25519_fe.h
… src/ballet/ed25519/avx/fd_ed25519_fe_avx.h
… src/ballet/ed25519/avx/fd_ed25519_fe_avx_inl.h
… src/ballet/ed25519/avx/fd_ed25519_ge.c
… src/ballet/ed25519/avx512/fd_f25519.h
… src/ballet/ed25519/avx512/fd_r43x6_ge.c
… src/ballet/ed25519/avx512/fd_r43x6_ge.h
… src/ballet/ed25519/avx512/fd_r43x6_ge_dmul_table.c
… src/ballet/ed25519/avx512/fd_r43x6_ge_smul_table.c
… src/ballet/ed25519/avx512/test_r43x6.c
… src/ballet/ed25519/fd_curve25519.c
… src/ballet/ed25519/fd_curve25519.h
… src/ballet/ed25519/fd_curve25519_scalar.c
… src/ballet/ed25519/fd_curve25519_scalar.h
… src/ballet/ed25519/fd_curve25519_secure.c
… src/ballet/ed25519/fd_curve25519_tables.c
… src/ballet/ed25519/fd_ed25519.h
… src/ballet/ed25519/fd_ed25519_fe.c
… src/ballet/ed25519/fd_ed25519_ge.c
… src/ballet/ed25519/fd_ed25519_private.h
… src/ballet/ed25519/fd_ed25519_user.c
… src/ballet/ed25519/fd_f25519.c
… src/ballet/ed25519/fd_f25519.h
… src/ballet/ed25519/fd_ristretto255.c
… src/ballet/ed25519/fd_ristretto255.h
… src/ballet/ed25519/fd_x25519.c